Split compare detail content sections
